requirements.txt
python-telegram-bot>=21.6,<22
python-dotenv>=1.0.1,<2
SQLAlchemy>=2.0,<3.0
asyncpg>=0.29,<1
httpx>=0.27,<1
